using float instead of Decimal

+++ Miner.py
@@ -1,5 +1,4 @@
 from Queue import Queue
-from decimal import Decimal
 from threading import Thread
 from time import time
 
@@ -29,7 +28,7 @@ class Miner(object):
 
 	def update_rate(self, now, iterations, t, targetQ, rate_divisor=1000):
 		self.rate = int((iterations / t) / rate_divisor)
-		self.rate = Decimal(self.rate) / 1000
+		self.rate = float(self.rate) / 1000
 		if self.accept_hist:
 			LAH = self.accept_hist.pop()
 			if LAH[1] != self.share_count[1]:
@@ -38,7 +37,7 @@ class Miner(object):
 		while (self.accept_hist[0][0] < now - self.options.estimate):
 			self.accept_hist.pop(0)
 		new_accept = self.share_count[1] - self.accept_hist[0][1]
-		self.estimated_rate = Decimal(new_accept) * (targetQ) / min(int(now - self.start_time), self.options.estimate) / 1000
-		self.estimated_rate = Decimal(self.estimated_rate) / 1000
+		self.estimated_rate = float(new_accept) * (targetQ) / min(int(now - self.start_time), self.options.estimate) / 1000
+		self.estimated_rate = float(self.estimated_rate) / 1000
 
 		self.switch.status_updated(self)
